# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample API",
        "description": "https://order-express-rtcb-git-master-shohruzbekcoder007s-projects.vercel.app/ and http://localhost:8085",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8085/api",
            "description": "local server"
        },
        {
            "url": "https://order-express-rtcb-git-master-shohruzbekcoder007s-projects.vercel.app/api",
            "description": "render server"
        }
    ],
    "paths": {
...
        "/foodtype/foodtypes": {
            "post": {
                "tags": [
                    "Foodtype"
                ],
                "summary": "Add a new foodtype",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name",
                                    "description"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Foodtype created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foodtype"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "summary": "Get all foodtypes",
                "tags": [
                    "Foodtype"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of foodtypes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Foodtype"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/foodtype/foodtype/{id}": {
            "get": {
                "summary": "Get a foodtype by ID",
                "tags": [
                    "Foodtype"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foodtype found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foodtype"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Foodtype not found"
                    }
                }
            }
        },
        "/foodtype/foodtypes/{id}": {
            "put": {
                "summary": "Update a foodtype",
                "tags": [
                    "Foodtype"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Foodtype updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Foodtype"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "summary": "Delete a foodtype",
                "tags": [
                    "Foodtype"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Foodtype deleted successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/food/foods": {
            "post": {
                "tags": [
                    "Food"
                ],
                "summary": "Add a new food",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "foodtypeId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "image",
                                    "foodtypeId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Food created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Food"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "get": {
                "summary": "Get all foods",
                "tags": [
                    "Food"
                ],
                "responses": {
                    "200": {
                        "description": "A list of foods",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Food"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/food/food/:id": {
            "get": {
                "summary": "Get a food by ID",
                "tags": [
                    "Food"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Food"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Food not found"
                    }
                }
            },
            "put": {
                "summary": "Update a food",
                "tags": [
                    "Food"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "foodtypeId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Food updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Food"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "summary": "Delete a food",
                "tags": [
                    "Food"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Food deleted successfully"
                    },
                    "404": {
                        "description": "Food not found"
                    }
                }
            }
        },
        "/food/foodtype/:foodtypeId": {
            "get": {
                "summary": "Get foods by foodtypeId",
                "tags": [
                    "Food"
                ],
                "parameters": [
                    {
                        "name": "foodtypeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foods found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Food"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No foods found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Foodtype": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
            "Food": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "integer"
                    },
                    "image": {
                        "type": "string"
                    },
                    "foodtypeId": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
...
}